BFD-3665
refactored enum and update libraries and created constants for the urls
