+++ apps/bfd-pipeline/bfd-pipeline-shared-utils/src/main/java/gov/cms/bfd/pipeline/sharedutils/SamhsaUtil.java
@@ -27,8 +27,8 @@ import gov.cms.bfd.pipeline.sharedutils.adapters.SamhsaInpatientAdapter;
 import gov.cms.bfd.pipeline.sharedutils.adapters.SamhsaOutpatientAdapter;
 import gov.cms.bfd.pipeline.sharedutils.adapters.SamhsaSnfAdapter;
 import gov.cms.bfd.pipeline.sharedutils.model.SamhsaEntry;
-import gov.cms.bfd.pipeline.sharedutils.model.TagCode;
 import gov.cms.bfd.pipeline.sharedutils.model.TagDetails;
+import gov.cms.bfd.sharedutils.TagCode;
 import jakarta.persistence.EntityManager;
 import java.io.IOException;
 import java.io.InputStream;
